[WEIGHTED_AVG] count implementation
