+++ src_cpp/common/common_definitions.h
@@ -1,4 +1,8 @@
 #pragma once
+
+#include "nifpp.h"
+#include <Logger.h>
+
 namespace nerlnet
 {
