+++ src_cpp/common/nerlWorker.h
@@ -1,5 +1,6 @@
 #pragma once
 
+#include <Logger.h>
 #include "nerlWorkerFunc.h"
 #include "nerlLayer.h"
 
@@ -22,6 +23,9 @@ namespace nerlnet
     int get_loss_method() { return _loss_method; };
     int get_distributed_system_type() { return _distributed_system_type; };
 
+    std::vector<int> get_distributed_system_train_labels_count() {LogError<<"Distributed System Weighted Avg count label is unsupported";
+                                                                                          throw("Distributed System Weighted Avg count label is unsupported");} // counts the number of each label appears in the training set for weighted average
+
     protected:
 
     std::shared_ptr<NerlLayer> _nerl_layers_linked_list;
@@ -33,6 +37,8 @@ namespace nerlnet
     int _optimizer_type;
     int _loss_method;
     std::string _distributed_system_args_str;
+    std::shared_ptr<std::vector<int>> _train_labels_count; // accumulates the number of each label in the training set
+
 
     private:
