+++ src_cpp/common/worker_definitions_ag.h
@@ -17,6 +17,6 @@ enum ModelTypeEnum{MODEL_TYPE_NN=0,MODEL_TYPE_APPROXIMATION=1,MODEL_TYPE_CLASSIF
 enum OptimizerEnum{OPTIMIZER_GD=0,OPTIMIZER_CGD=1,OPTIMIZER_SGD=2,OPTIMIZER_QUASINEUTON=3,OPTIMIZER_LVM=4,OPTIMIZER_ADAM=5};
 enum LossMethodEnum{LOSS_METHOD_SSE=1,LOSS_METHOD_MSE=2,LOSS_METHOD_NSE=3,LOSS_METHOD_MINKOWSKIE=4,LOSS_METHOD_WSE=5,LOSS_METHOD_CEE=6};
 enum InfraTypeEnum{INFRA_TYPE_OPENNN=0,INFRA_TYPE_WOLFENGINE=1};
-enum WorkerDistributedSystemTypeEnum{WORKER_DISTRIBUTED_SYSTEM_TYPE_NONE=0,WORKER_DISTRIBUTED_SYSTEM_TYPE_FEDCLIENTAVG=1,WORKER_DISTRIBUTED_SYSTEM_TYPE_FEDSERVERAVG=2};
+enum WorkerDistributedSystemTypeEnum{WORKER_DISTRIBUTED_SYSTEM_TYPE_NONE=0,WORKER_DISTRIBUTED_SYSTEM_TYPE_FEDCLIENTAVG=1,WORKER_DISTRIBUTED_SYSTEM_TYPE_FEDSERVERAVG=2,WORKER_DISTRIBUTED_SYSTEM_TYPE_FEDCLIENTWEGITHEDAVGCLASSIFICATION=3,WORKER_DISTRIBUTED_SYSTEM_TYPE_FEDSERVERWEGITHEDAVGCLASSIFICATION=4,WORKER_DISTRIBUTED_SYSTEM_TYPE_FEDCLIENTAE=5,WORKER_DISTRIBUTED_SYSTEM_TYPE_FEDSERVERAE=6};
 
 } // namespace nerlnet
