+++ src_erl/NerlnetApp/src/worker_definitions_ag.hrl
@@ -34,14 +34,26 @@
 -define(DC_DISTRIBUTED_SYSTEM_TYPE_NONE_KEY_ATOM,none).
 -define(DC_DISTRIBUTED_SYSTEM_TYPE_FEDCLIENTAVG_KEY_ATOM,fedClientAvg).
 -define(DC_DISTRIBUTED_SYSTEM_TYPE_FEDSERVERAVG_KEY_ATOM,fedServerAvg).
+-define(DC_DISTRIBUTED_SYSTEM_TYPE_FEDCLIENTWEGITHEDAVGCLASSIFICATION_KEY_ATOM,fedClientWegithedAvgClassification).
+-define(DC_DISTRIBUTED_SYSTEM_TYPE_FEDSERVERWEGITHEDAVGCLASSIFICATION_KEY_ATOM,fedServerWegithedAvgClassification).
+-define(DC_DISTRIBUTED_SYSTEM_TYPE_FEDCLIENTAE_KEY_ATOM,fedClientAE).
+-define(DC_DISTRIBUTED_SYSTEM_TYPE_FEDSERVERAE_KEY_ATOM,fedServerAE).
 
 -define(DC_DISTRIBUTED_SYSTEM_TYPE_NONE_IDX_STR,"0").
 -define(DC_DISTRIBUTED_SYSTEM_TYPE_FEDCLIENTAVG_IDX_STR,"1").
 -define(DC_DISTRIBUTED_SYSTEM_TYPE_FEDSERVERAVG_IDX_STR,"2").
+-define(DC_DISTRIBUTED_SYSTEM_TYPE_FEDCLIENTWEGITHEDAVGCLASSIFICATION_IDX_STR,"3").
+-define(DC_DISTRIBUTED_SYSTEM_TYPE_FEDSERVERWEGITHEDAVGCLASSIFICATION_IDX_STR,"4").
+-define(DC_DISTRIBUTED_SYSTEM_TYPE_FEDCLIENTAE_IDX_STR,"5").
+-define(DC_DISTRIBUTED_SYSTEM_TYPE_FEDSERVERAE_IDX_STR,"6").
 
 -define(DC_DISTRIBUTED_SYSTEM_TYPE_NONE_IDX,0).
 -define(DC_DISTRIBUTED_SYSTEM_TYPE_FEDCLIENTAVG_IDX,1).
 -define(DC_DISTRIBUTED_SYSTEM_TYPE_FEDSERVERAVG_IDX,2).
+-define(DC_DISTRIBUTED_SYSTEM_TYPE_FEDCLIENTWEGITHEDAVGCLASSIFICATION_IDX,3).
+-define(DC_DISTRIBUTED_SYSTEM_TYPE_FEDSERVERWEGITHEDAVGCLASSIFICATION_IDX,4).
+-define(DC_DISTRIBUTED_SYSTEM_TYPE_FEDCLIENTAE_IDX,5).
+-define(DC_DISTRIBUTED_SYSTEM_TYPE_FEDSERVERAE_IDX,6).
 
 -define(DC_INFRA_TYPE_OPENNN_KEY_ATOM,opennn).
 -define(DC_INFRA_TYPE_WOLFENGINE_KEY_ATOM,wolfengine).
